requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
fastapi-users[sqlalchemy]>=14.0.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.2.0
httpx>=0.27.0
email-validator>=2.2.0
huggingface_hub>=0.24.0
PyYAML>=6.0
pgvector>=0.3.0
# Document parsing for knowledge base uploads
python-docx>=1.0.0
openpyxl>=3.1.0
python-pptx>=0.6.0
PyPDF2>=3.0.0
odfpy>=1.4.0
